Rewrite join-game usecase
