+++ love_letter/usecase/__init__.py
@@ -4,41 +4,9 @@ from typing import Dict, Union
 from love_letter.models import Game, GuessCard, Player, ToSomeoneCard
 from love_letter.repository import create_default_repository
 from love_letter.usecase.create_game import CreateGame
+from love_letter.usecase.join_game import JoinGame
 
 game_repository = create_default_repository()
-
-
-class JoinGameInput:
-    game_id: str
-    player_id: str
-
-
-class JoinGameOutput:
-    success: bool
-
-
-class JoinGame:
-    def execute(self, input: JoinGameInput, output: JoinGameOutput):
-        try:
-            game = game_repository.get(input.game_id)
-            game.join(Player(input.player_id))
-            game_repository.save_or_update(game)
-            output.success = True
-        except BaseException as e:
-            traceback.print_exception(e)
-            output.success = False
-
-    @classmethod
-    def output(cls) -> JoinGameOutput:
-        output = JoinGameOutput()
-        return output
-
-    @classmethod
-    def input(cls, game_id: str, player_id: str) -> JoinGameInput:
-        input = JoinGameInput()
-        input.game_id = game_id
-        input.player_id = player_id
-        return input
 
 
 class StartGameInput:
